Refactor hints
No need to duplicate this logic.
