+++ server/src/hovers/index.ts
@@ -1,7 +1,11 @@
 import { Hover } from "vscode-languageserver/node";
 
 import { get } from "../apiClient";
-import { urlFor } from "../prefab";
+import { methodAtPosition } from "../documentAnnotations";
+import {
+  filterForMissingKeys as defaultFilterForMissingKeys,
+  urlFor,
+} from "../prefab";
 import type { HoverAnalyzer, HoverAnalyzerArgs } from "../types";
 import evaluations from "./evaluations";
 import linkTitle from "./linkTitle";
@@ -15,9 +19,29 @@ type Dependencies = {
 };
 
 export const runAllHovers = async (
-  args: HoverAnalyzerArgs & Dependencies
+  args: Omit<HoverAnalyzerArgs, "method"> & Dependencies
 ): Promise<Hover | null> => {
-  const allHovers = await Promise.all(hovers.map((hover) => hover(args)));
+  const { filterForMissingKeys, log, document, position } = args;
+
+  const method = methodAtPosition(document, position);
+
+  if (!method) {
+    log("Hover", "No method found at position");
+    return null;
+  }
+
+  const missingKeyMethods = await (
+    filterForMissingKeys ?? defaultFilterForMissingKeys
+  )([method]);
+
+  if (missingKeyMethods.length > 0) {
+    log("Hover", "Key does not exist");
+    return null;
+  }
+
+  const allHovers = await Promise.all(
+    hovers.map((hover) => hover({ ...args, method }))
+  );
 
   const filteredHovers: Hover[] = allHovers.filter(
     (hover): hover is Hover => hover !== null
